refactor(api): address review + apply architecture improvements
PR review fixes (#1073):
- walkProperty: surface element schema name on `ref` for `array<$ref>`
  body fields (and the equivalent on path/query params). Without this,
  the schemas-referenced footer silently omitted the array element type.
- Stale `parseOpenApi` docstring updated to match the synthesize-and-
  debug-log behavior.
- Path-item + operation-level params now dedup by `(name, in)` with
  operation-level wins (per OpenAPI 3.x).
- `isStreamingResponse` now matches any 2xx with `text/event-stream`
  (was 200-only).
- Tighten the changeset to a single user-facing sentence.

Architecture improvements (from /improve-codebase-architecture):
- Unify HTTP fetch: `docsClient.fetchSpec(slug, {format?})` now
  accepts `'yaml' | 'json'`. `openapi/get.ts` drops its hand-rolled
  fetch + URL/timeout/headers constants and goes through docsClient,
  so the docs-endpoint URL + Vercel bypass token + 404 handling
  live in one module.
- Thread component schemas through `ParsedSpec.schemas` instead of
  re-parsing YAML in `--schema <name>`. Drops `lookupComponentSchema`
  / `listComponentSchemas` exports and the `yaml` import from
  parser.ts. One YAML parse per `--schema` invocation, no second
  parse pass.
- Split parser.ts (780 lines) into three files: `parser.ts` keeps
  types + capability + URL helpers + parseOpenApi orchestration +
  loaders; `extractors.ts` holds the per-aspect extractors
  (parameters / requestBody / responses / security); `internal.ts`
  holds shared shape-coercion helpers (asObject/asString/asArray)
  and the OpenAPI-shape utilities (schemaRefName / describeType /
  summarizeInlineObject). Dependency direction: parser → extractors
  → internal.
