+++ packages/@sanity/cli/src/api/parser.ts
@@ -1,9 +1,15 @@
 import {subdebug} from '@sanity/cli-core'
 import {validate} from '@scalar/openapi-parser'
 import {type OpenAPIV3, type OpenAPIV3_1} from '@scalar/openapi-types'
-import {parse as parseYaml} from 'yaml'
 
 import {fetchSpec, fetchSpecIndex, type OpenApiSpecIndexEntry} from './docsClient.js'
+import {
+  extractParameters,
+  extractRequestBody,
+  extractResponses,
+  extractSecurity,
+  isStreamingResponse,
+} from './extractors.js'
 
 const debug = subdebug('api:parser')
 
@@ -89,7 +95,7 @@ export interface ParsedResponse {
   ref?: string
 }
 
-interface SecurityScheme {
+export interface SecurityScheme {
   /** Normalized scheme name, e.g. `BearerAuth`. */
   scheme: string
 }
@@ -120,6 +126,12 @@ export interface ParsedOperation {
 export interface ParsedSpec {
   description: string
   operations: ParsedOperation[]
+  /**
+   * `components.schemas` keyed by name. Empty when the spec doesn't
+   * declare any. Used by `sanity api spec --schema <name>` to follow
+   * `$ref` pointers without re-parsing the YAML.
+   */
+  schemas: Record<string, unknown>
   slug: string
   title: string
   /** `info.version` (may be a meaningless semver — the endpoint version comes from the server URL). */
@@ -210,404 +222,6 @@ function extractServerPathSegment(serverUrlWithPlaceholders: string): string {
 }
 
 /* ---------------------------------------------------------------------- *
- *  $ref helpers                                                           *
- * ---------------------------------------------------------------------- */
-
-const SCHEMA_REF_PREFIX = '#/components/schemas/'
-const PARAMETER_REF_PREFIX = '#/components/parameters/'
-
-/**
- * Extract the schema name from `#/components/schemas/Foo`. Returns ''
- * for non-local or malformed refs — we don't follow remote refs.
- */
-function schemaRefName(ref: string): string {
-  return ref.startsWith(SCHEMA_REF_PREFIX) ? ref.slice(SCHEMA_REF_PREFIX.length) : ''
-}
-
-/**
- * Resolve a `#/components/parameters/<name>` ref against the spec root.
- * Unlike schema refs (link-not-resolve), parameter refs MUST resolve —
- * path params especially: without them, the URL template loses
- * placeholders. 8 of 22 Sanity specs use this pattern.
- */
-function resolveParameterRef(
-  ref: string,
-  root: OpenAPIV3_1.Document,
-): OpenAPIV3.ParameterObject | OpenAPIV3_1.ParameterObject | null {
-  if (!ref.startsWith(PARAMETER_REF_PREFIX)) return null
-  const name = ref.slice(PARAMETER_REF_PREFIX.length)
-  const params = root.components?.parameters
-  const target = params?.[name]
-  if (!target || '$ref' in target) return null
-  return target as OpenAPIV3.ParameterObject | OpenAPIV3_1.ParameterObject
-}
-
-/* ---------------------------------------------------------------------- *
- *  Schema type / summary helpers                                          *
- * ---------------------------------------------------------------------- */
-
-type SchemaLike = Record<string, unknown>
-
-function asObject(value: unknown): SchemaLike {
-  return value && typeof value === 'object' && !Array.isArray(value) ? (value as SchemaLike) : {}
-}
-
-function asString(value: unknown): string {
-  return typeof value === 'string' ? value : ''
-}
-
-function asArray<T>(value: unknown): T[] {
-  return Array.isArray(value) ? (value as T[]) : []
-}
-
-/**
- * One-word type label for table / help display. Refs collapse to the
- * schema name (`'Foo'`) so the output is immediately useful without
- * resolution. Composition (`allOf`/`oneOf`/`anyOf`) collapses to
- * `'object'`; the body walker handles composition itself.
- */
-function describeType(schema: SchemaLike): string {
-  const ref = asString(schema.$ref)
-  if (ref) return schemaRefName(ref) || 'unknown'
-  const t = asString(schema.type)
-  if (t) {
-    if (t === 'array') {
-      const inner = describeType(asObject(schema.items))
-      return inner === 'unknown' ? 'array' : `${inner}[]`
-    }
-    return t
-  }
-  if (Array.isArray(schema.enum) && schema.enum.length > 0) {
-    return typeof schema.enum[0] === 'number' ? 'number' : 'string'
-  }
-  if (Array.isArray(schema.allOf) || Array.isArray(schema.oneOf) || Array.isArray(schema.anyOf)) {
-    return 'object'
-  }
-  return 'unknown'
-}
-
-/** Compact `{ a, b, c }` summary of an inline object schema. */
-function summarizeInlineObject(properties: SchemaLike): string {
-  const names = Object.keys(properties)
-  if (names.length === 0) return ''
-  const head = names.slice(0, 6).join(', ')
-  return names.length > 6 ? `{ ${head}, … }` : `{ ${head} }`
-}
-
-/* ---------------------------------------------------------------------- *
- *  Parameter extraction                                                   *
- * ---------------------------------------------------------------------- */
-
-type ParameterObject = OpenAPIV3.ParameterObject | OpenAPIV3_1.ParameterObject
-type RefOrParameter = OpenAPIV3.ReferenceObject | ParameterObject
-
-/**
- * Resolve a (possibly $ref'd) parameter object and extract a
- * `ParsedParam` with full schema metadata. Returns null for malformed
- * or unresolvable params so callers can filter without crashing.
- */
-function parseParam(rawParam: RefOrParameter, root: OpenAPIV3_1.Document): ParsedParam | null {
-  const param: ParameterObject | null =
-    '$ref' in rawParam ? resolveParameterRef(rawParam.$ref, root) : rawParam
-  if (!param) return null
-
-  const location = param.in
-  if (!param.name || (location !== 'path' && location !== 'query' && location !== 'header')) {
-    return null
-  }
-
-  const schema = asObject(param.schema)
-  // OpenAPI lets `enum`/`default`/`example` live on either the param or its schema; check both.
-  const enumValues =
-    asArray<number | string>(schema.enum).length > 0
-      ? asArray<number | string>(schema.enum)
-      : asArray<number | string>((param as Record<string, unknown>).enum)
-
-  const parsed: ParsedParam = {
-    description: param.description || asString(schema.description),
-    in: location,
-    name: param.name,
-    required: location === 'path' ? true : param.required === true,
-    type: describeType(schema),
-  }
-
-  const ref = asString(schema.$ref)
-  if (ref) {
-    const refName = schemaRefName(ref)
-    if (refName) parsed.ref = refName
-  }
-  if (enumValues.length > 0) parsed.enum = enumValues
-  const defaultValue = schema.default ?? (param as Record<string, unknown>).default
-  if (defaultValue !== undefined) parsed.default = defaultValue
-  const exampleValue = schema.example ?? (param as Record<string, unknown>).example
-  if (exampleValue !== undefined) parsed.example = exampleValue
-  return parsed
-}
-
-/* ---------------------------------------------------------------------- *
- *  Request body extraction                                                *
- * ---------------------------------------------------------------------- */
-
-/**
- * Walk one property schema into a `ParsedBodyField`. Refs stop here:
- * the ref name lands on `ref`, recursion does not follow. Inline objects
- * (and arrays-of-objects) recurse one level deeper.
- */
-function walkProperty(name: string, rawSchema: unknown, isRequired: boolean): ParsedBodyField {
-  const schema = asObject(rawSchema)
-  const ref = asString(schema.$ref)
-  const refName = ref ? schemaRefName(ref) : ''
-
-  const field: ParsedBodyField = {
-    description: asString(schema.description),
-    fields: [],
-    name,
-    required: isRequired,
-    type: describeType(schema),
-  }
-  if (refName) field.ref = refName
-  const enumValues = asArray<number | string>(schema.enum)
-  if (enumValues.length > 0) field.enum = enumValues
-  if (schema.default !== undefined) field.default = schema.default
-
-  // Don't follow refs — link-not-resolve.
-  if (ref) return field
-
-  // Recurse into inline objects / arrays-of-objects.
-  const t = asString(schema.type)
-  if (t === 'object' || schema.properties) {
-    field.fields = walkProperties(schema)
-  } else if (t === 'array') {
-    const items = asObject(schema.items)
-    // `items: {$ref: ...}` is the common shape for arrays of refs.
-    // Surface the element schema name so the footer can drill into it.
-    const itemsRef = asString(items.$ref)
-    if (itemsRef) {
-      const itemsRefName = schemaRefName(itemsRef)
-      if (itemsRefName) field.ref = itemsRefName
-    } else if (asString(items.type) === 'object' || items.properties) {
-      field.fields = walkProperties(items)
-    }
-  }
-  return field
-}
-
-function walkProperties(schema: SchemaLike): ParsedBodyField[] {
-  const properties = asObject(schema.properties)
-  const required = new Set(asArray<string>(schema.required))
-  return Object.entries(properties).map(([propName, propSchema]) =>
-    walkProperty(propName, propSchema, required.has(propName)),
-  )
-}
-
-/**
- * Flatten `allOf` at a schema root by collecting inline properties from
- * every variant. `oneOf`/`anyOf` take the first variant for properties.
- * Refs at any variant slot land in `refs` for caller to surface as
- * follow-up pointers.
- */
-function flattenComposition(schema: SchemaLike): {
-  fields: ParsedBodyField[]
-  refs: string[]
-} {
-  const refs = new Set<string>()
-  const fieldsByName = new Map<string, ParsedBodyField>()
-  for (const f of walkProperties(schema)) fieldsByName.set(f.name, f)
-
-  const ownRef = asString(schema.$ref)
-  if (ownRef) {
-    const name = schemaRefName(ownRef)
-    if (name) refs.add(name)
-  }
-
-  for (const variant of asArray<SchemaLike>(schema.allOf)) {
-    const sub = flattenComposition(variant)
-    for (const f of sub.fields) {
-      if (!fieldsByName.has(f.name)) fieldsByName.set(f.name, f)
-    }
-    for (const r of sub.refs) refs.add(r)
-  }
-
-  for (const key of ['oneOf', 'anyOf'] as const) {
-    const variants = asArray<SchemaLike>(schema[key])
-    for (const [i, variant] of variants.entries()) {
-      const variantRef = asString(variant.$ref)
-      if (variantRef) {
-        const name = schemaRefName(variantRef)
-        if (name) refs.add(name)
-        continue
-      }
-      // Inline variants: use the first one's properties for a representative shape.
-      if (i === 0) {
-        const sub = flattenComposition(variant)
-        for (const f of sub.fields) {
-          if (!fieldsByName.has(f.name)) fieldsByName.set(f.name, f)
-        }
-        for (const r of sub.refs) refs.add(r)
-      }
-    }
-  }
-
-  return {fields: [...fieldsByName.values()], refs: [...refs]}
-}
-
-function pickContentType(content: SchemaLike): string {
-  const keys = Object.keys(content)
-  if (keys.length === 0) return ''
-  if ('application/json' in content) return 'application/json'
-  return keys[0]
-}
-
-function isJsonContentType(contentType: string): boolean {
-  return contentType === 'application/json' || contentType.endsWith('+json')
-}
-
-function summarizeBodySchema(schema: SchemaLike, refs: string[]): string {
-  const inline = summarizeInlineObject(asObject(schema.properties))
-  if (inline) return inline
-  if (refs.length > 0) return refs.length === 1 ? refs[0] : `oneOf(${refs.join(', ')})`
-  return describeType(schema)
-}
-
-type RefOrRequestBody = OpenAPIV3.ReferenceObject | OpenAPIV3.RequestBodyObject
-
-function parseRequestBody(opRaw: OperationObject): ParsedRequestBody | null {
-  const body = opRaw.requestBody as RefOrRequestBody | undefined
-  if (!body) return null
-  // Body-level $refs are rare; specs we own don't use them. Treat as no body.
-  if ('$ref' in body) return null
-
-  const content = asObject(body.content)
-  if (Object.keys(content).length === 0) return null
-
-  const contentType = pickContentType(content)
-  const mediaTypeObj = asObject(content[contentType])
-  const schema = asObject(mediaTypeObj.schema)
-
-  if (!isJsonContentType(contentType)) {
-    return {
-      contentType,
-      fields: [],
-      refs: [],
-      required: body.required === true,
-      schemaSummary: `<${contentType}>`,
-    }
-  }
-
-  const {fields, refs} = flattenComposition(schema)
-  return {
-    contentType,
-    fields,
-    refs,
-    required: body.required === true,
-    schemaSummary: summarizeBodySchema(schema, refs),
-  }
-}
-
-/* ---------------------------------------------------------------------- *
- *  Response extraction                                                    *
- * ---------------------------------------------------------------------- */
-
-function parseResponses(opRaw: OperationObject): ParsedResponse[] {
-  const responses = asObject(opRaw.responses)
-  const out: ParsedResponse[] = []
-  for (const [statusKey, rawResponse] of Object.entries(responses)) {
-    const status = statusKey === 'default' ? 0 : Number.parseInt(statusKey, 10)
-    if (!Number.isFinite(status) && statusKey !== 'default') continue
-    const response = asObject(rawResponse)
-    // Skip response refs — uncommon in our specs, and the schema lookup
-    // covers the use case if anyone needs it.
-    if (asString(response.$ref)) continue
-
-    const content = asObject(response.content)
-    const contentType = pickContentType(content)
-    const mediaObj = asObject(content[contentType])
-    const schema = asObject(mediaObj.schema)
-
-    let schemaSummary = ''
-    let ref: string | undefined
-    if (contentType) {
-      const schemaRef = asString(schema.$ref)
-      if (schemaRef) {
-        const name = schemaRefName(schemaRef)
-        if (name) {
-          ref = name
-          schemaSummary = name
-        }
-      } else if (isJsonContentType(contentType)) {
-        schemaSummary =
-          summarizeInlineObject(asObject(schema.properties)) ||
-          (Object.keys(schema).length > 0 ? describeType(schema) : '')
-      } else {
-        schemaSummary = `<${contentType}>`
-      }
-    }
-
-    const entry: ParsedResponse = {contentType, schemaSummary, status}
-    if (ref) entry.ref = ref
-    out.push(entry)
-  }
-  return out.toSorted((a, b) => a.status - b.status)
-}
-
-function isStreamingResponse(responses: ParsedResponse[]): boolean {
-  return responses.some((r) => r.status === 200 && r.contentType === 'text/event-stream')
-}
-
-/* ---------------------------------------------------------------------- *
- *  Security extraction                                                    *
- * ---------------------------------------------------------------------- */
-
-/**
- * Specs inconsistently spell `BearerAuth` / `bearerAuth`. Same scheme;
- * capitalize the first letter for uniform display.
- */
-function normalizeSchemeName(name: string): string {
-  if (!name) return ''
-  return name[0].toUpperCase() + name.slice(1)
-}
-
-function parseSecurity(opRaw: OperationObject, root: OpenAPIV3_1.Document): SecurityScheme[] {
-  const opSecurity = asArray<SchemaLike>(opRaw.security)
-  const source = opSecurity.length > 0 ? opSecurity : asArray<SchemaLike>(root.security)
-  const schemes = new Set<string>()
-  for (const requirement of source) {
-    for (const key of Object.keys(requirement)) schemes.add(normalizeSchemeName(key))
-  }
-  return [...schemes].map((scheme) => ({scheme}))
-}
-
-/* ---------------------------------------------------------------------- *
- *  Component schema lookup (for `--schema <name>`)                        *
- * ---------------------------------------------------------------------- */
-
-/**
- * Return the raw `components.schemas.<name>` object from a spec YAML,
- * or `null` if the schema doesn't exist. Used by the `--schema` flag
- * on `api spec` to let agents follow `$ref` pointers surfaced in
- * operation output.
- *
- * Reads YAML directly rather than going through validate() — this is a
- * synchronous lookup over the raw document, no extraction logic needed.
- */
-export function lookupComponentSchema(yaml: string, name: string): unknown {
-  const doc = parseYaml(yaml)
-  if (!doc || typeof doc !== 'object') return null
-  const components = asObject((doc as SchemaLike).components)
-  const schemas = asObject(components.schemas)
-  return name in schemas ? schemas[name] : null
-}
-
-/** List all `components.schemas` keys defined by a spec. */
-export function listComponentSchemas(yaml: string): string[] {
-  const doc = parseYaml(yaml)
-  if (!doc || typeof doc !== 'object') return []
-  const components = asObject((doc as SchemaLike).components)
-  const schemas = asObject(components.schemas)
-  return Object.keys(schemas)
-}
-
-/* ---------------------------------------------------------------------- *
  *  OpenAPI parser                                                         *
  * ---------------------------------------------------------------------- */
 
@@ -624,7 +238,9 @@ type PathItemObject = OpenAPIV3.PathItemObject | OpenAPIV3_1.PathItemObject
  * etc.) are not fatal — surfaces whatever it can read. Only hard parse
  * failures (`specification` missing entirely) throw.
  *
- * Operations without an `operationId` are skipped and debug-logged.
+ * Operations without an `operationId` get a deterministic synthesized
+ * id from method + path (see `synthesizeOperationId`); the synthesis
+ * is debug-logged.
  */
 export async function parseOpenApi(slug: string, yaml: string): Promise<ParsedSpec> {
   const result = await validate(yaml)
@@ -658,7 +274,6 @@ export async function parseOpenApi(slug: string, yaml: string): Promise<ParsedSp
   for (const [rawPath, pathItem] of Object.entries(doc.paths ?? {})) {
     if (!pathItem) continue
     const pathItemObj = pathItem as PathItemObject
-    const commonParams = pathItemObj.parameters
 
     for (const method of HTTP_METHODS) {
       const op = pathItemObj[method as HttpMethod] as OperationObject | undefined
@@ -674,11 +289,12 @@ export async function parseOpenApi(slug: string, yaml: string): Promise<ParsedSp
         debug(`synthesized operationId "${operationId}" for ${methodUpper} ${rawPath}`)
       }
 
-      const allParams = ([...(commonParams ?? []), ...(op.parameters ?? [])] as RefOrParameter[])
-        .map((raw) => parseParam(raw, doc))
-        .filter((p): p is ParsedParam => p !== null)
-
-      const responses = parseResponses(op)
+      const {headerParams, pathParams, queryParams} = extractParameters(
+        pathItemObj.parameters,
+        op.parameters,
+        doc,
+      )
+      const responses = extractResponses(op)
       const opPath = rawPath.replace(/^\/+/, '')
       const combined = serverPathSegment ? `${serverPathSegment}/${opPath}` : opPath
 
@@ -686,17 +302,17 @@ export async function parseOpenApi(slug: string, yaml: string): Promise<ParsedSp
         capability: classifyCapability(methodUpper),
         description: op.description ?? '',
         endpoint: toUrlPatternForm(combined),
-        headerParams: allParams.filter((p) => p.in === 'header'),
+        headerParams,
         isStreaming: isStreamingResponse(responses),
         method: methodUpper,
         operationId,
         optionalQueryParams,
         path: rawPath,
-        pathParams: allParams.filter((p) => p.in === 'path'),
-        queryParams: allParams.filter((p) => p.in === 'query'),
-        requestBody: parseRequestBody(op),
+        pathParams,
+        queryParams,
+        requestBody: extractRequestBody(op),
         responses,
-        security: parseSecurity(op, doc),
+        security: extractSecurity(op, doc),
         summary: op.summary ?? '',
       })
     }
@@ -705,6 +321,7 @@ export async function parseOpenApi(slug: string, yaml: string): Promise<ParsedSp
   return {
     description: doc.info?.description ?? '',
     operations,
+    schemas: (doc.components?.schemas as Record<string, unknown> | undefined) ?? {},
     slug,
     title: doc.info?.title || slug,
     version: doc.info?.version ?? '',
@@ -800,7 +417,7 @@ function sortOperations(entries: OperationIndexEntry[]): OperationIndexEntry[] {
 /**
  * Result of a single-spec fetch. The raw YAML is kept alongside the
  * parsed form so the spec command can pass it straight through for
- * `--format=openapi` and use it for `--schema <name>` lookups.
+ * `--format=openapi`.
  */
 interface LoadedSpec {
   index: OpenApiSpecIndexEntry
